requirements.txt
langchain
langchain_community
streamlit
openai
gtts
pygame
requests
python-dotenv